+++ hadoop-mapreduce/src/test/java/org/apache/accumulo/hadoop/its/mapreduce/RowHashIT.java
@@ -27,7 +27,9 @@ import java.time.Duration;
 import java.util.Base64;
 import java.util.Collections;
 import java.util.Map.Entry;
+import java.util.Properties;
 
+import org.apache.accumulo.core.cli.ClientOpts;
 import org.apache.accumulo.core.client.Accumulo;
 import org.apache.accumulo.core.client.AccumuloClient;
 import org.apache.accumulo.core.client.AccumuloException;
@@ -38,13 +40,17 @@ import org.apache.accumulo.core.client.MutationsRejectedException;
 import org.apache.accumulo.core.client.Scanner;
 import org.apache.accumulo.core.client.TableExistsException;
 import org.apache.accumulo.core.client.TableNotFoundException;
+import org.apache.accumulo.core.client.admin.DelegationTokenConfig;
+import org.apache.accumulo.core.client.security.tokens.AuthenticationToken;
+import org.apache.accumulo.core.client.security.tokens.KerberosToken;
+import org.apache.accumulo.core.conf.ClientProperty;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Mutation;
 import org.apache.accumulo.core.data.Value;
 import org.apache.accumulo.core.security.Authorizations;
+import org.apache.accumulo.core.security.SystemPermission;
 import org.apache.accumulo.hadoop.mapreduce.AccumuloInputFormat;
 import org.apache.accumulo.hadoop.mapreduce.AccumuloOutputFormat;
-import org.apache.accumulo.hadoopImpl.mapreduce.lib.MapReduceClientOpts;
 import org.apache.accumulo.miniclusterImpl.MiniAccumuloClusterImpl;
 import org.apache.accumulo.test.functional.ConfigurableMacBase;
 import org.apache.hadoop.conf.Configuration;
@@ -53,6 +59,7 @@ import org.apache.hadoop.io.MD5Hash;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.Mapper;
+import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.hadoop.util.Tool;
 import org.apache.hadoop.util.ToolRunner;
 import org.junit.jupiter.api.Test;
@@ -131,11 +138,63 @@ public class RowHashIT extends ConfigurableMacBase {
       public void setup(Context job) {}
     }
 
-    public class Opts extends MapReduceClientOpts {
+    public class Opts extends ClientOpts {
       @Parameter(names = "--column", required = true)
       String column;
       @Parameter(names = {"-t", "--table"}, required = true, description = "table to use")
       String tableName;
+
+      @Override
+      public Properties getClientProps() {
+        Properties props = super.getClientProps();
+        // For MapReduce, Kerberos credentials don't make it to the Mappers and Reducers,
+        // so we need to request a delegation token and use that instead.
+        AuthenticationToken authToken = ClientProperty.getAuthenticationToken(props);
+        if (authToken instanceof KerberosToken) {
+          log.info("Received KerberosToken, fetching DelegationToken for MapReduce");
+          final KerberosToken krbToken = (KerberosToken) authToken;
+
+          try {
+            UserGroupInformation user = UserGroupInformation.getCurrentUser();
+            if (!user.hasKerberosCredentials()) {
+              throw new IllegalStateException("Expected current user to have Kerberos credentials");
+            }
+
+            String newPrincipal = user.getUserName();
+            log.info("Obtaining delegation token for {}", newPrincipal);
+
+            try (AccumuloClient client =
+                Accumulo.newClient().from(props).as(newPrincipal, krbToken).build()) {
+
+              // Do the explicit check to see if the user has the permission to get a delegation
+              // token
+              if (!client.securityOperations().hasSystemPermission(client.whoami(),
+                  SystemPermission.OBTAIN_DELEGATION_TOKEN)) {
+                log.error(
+                    "{} doesn't have the {} SystemPermission necessary to obtain a delegation"
+                        + " token. MapReduce tasks cannot automatically use the client's"
+                        + " credentials on remote servers. Delegation tokens provide a means to run"
+                        + " MapReduce without distributing the user's credentials.",
+                    user.getUserName(), SystemPermission.OBTAIN_DELEGATION_TOKEN.name());
+                throw new IllegalStateException(
+                    client.whoami() + " does not have permission to obtain a delegation token");
+              }
+
+              // Get the delegation token from Accumulo
+              AuthenticationToken token =
+                  client.securityOperations().getDelegationToken(new DelegationTokenConfig());
+
+              props.setProperty(ClientProperty.AUTH_PRINCIPAL.getKey(), newPrincipal);
+              ClientProperty.setAuthenticationToken(props, token);
+            }
+          } catch (IOException | AccumuloException | AccumuloSecurityException e) {
+            final String msg = "Failed to acquire DelegationToken for use with MapReduce";
+            log.error(msg, e);
+            throw new RuntimeException(msg, e);
+          }
+        }
+        return props;
+      }
     }
 
     @Override
